collector/requirements.txt
azure-functions
azure-storage-blob
requests
pandas
pyarrow
datetime